move timesheet component logic to timesheet card component
